auth_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
click==8.2.1
databases==0.9.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
psycopg2==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
